refactor(numbergenerator): NumberGenerator 인터페이스 및 이를 구현한 클래스 패키지 생성
