+++ src/main/java/domain/RacingGame.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.numbergenerator.NumberGenerator;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
