+++ src/test/java/domain/RacingGameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.numbergenerator.RandomNumberGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
